+++ MKAgent/src/main/java/org/AIandGames/mancalabot/GameTreeNode.java
@@ -98,7 +98,7 @@ public class GameTreeNode {
         hs.add(new MKPointDifference(this));
         hs.add(new RightMostPot(this));
         hs.add(new NumberOfEmptyPots(this));
-        hs.add(new Steal(this));
+        hs.add(new CumulativeSteal(this));
         hs.add(new FifteenInPot(this));
         hs.add(new RepeatMoveAvailable(this));
