+++ MKAgent/src/main/java/org/AIandGames/mancalabot/Heuristics/HeuristicWeightings.java
@@ -17,7 +17,7 @@ public class HeuristicWeightings {
         weightings.put(Heuristics.MK_POINT_DIFFERENCE, 1.0);
         weightings.put(Heuristics.RIGHT_MOST_POT, 0.00001);
         weightings.put(Heuristics.NUMBER_OF_EMPTY_POTS, 0.2);
-        weightings.put(Heuristics.STEAL, 0.6);
+        weightings.put(Heuristics.CUMULATIVE_STEAL, 0.6);
         weightings.put(Heuristics.FIFTEEN_IN_POT, 2.0);
         weightings.put(Heuristics.REPEAT_MOVE_AVAILABLE, 0.1);
     }
